use hostname_validator for rfc1123 domain verification
