+++ pubky-homeserver/Cargo.toml
@@ -47,6 +47,7 @@ page_size = "0.6.0"
 pkarr-republisher = { version = "0.1.0", path = "../pkarr-republisher" }
 thiserror = "2.0.12"
 dirs = "6.0.0"
+hostname-validator = "1.1.1"
 
 [dev-dependencies]
 tempfile = "3.10.1"
