+++ pubky-homeserver/src/data_directory/mod.rs
@@ -1,7 +1,6 @@
 mod config_toml;
 mod data_dir;
 mod domain_port;
-mod validate_domain;
 
 pub use config_toml::{ConfigReadError, ConfigToml};
 pub use data_dir::DataDir;
